Feat: tied up components & refactor
