+++ tailwind.config.cjs
@@ -6,7 +6,16 @@ module.exports = {
     "./src/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
-    extend: {},
+    extend: {
+      colors: {
+        hashBlue: '#487EB0',
+        lightGray:'#EEEDED',
+        themeColor: '#EEEDED',
+        searchColor: '#F6F4F4',
+        buttonColor: '#487EBO',
+        tableHeader: '#D9D9D9'
+      }
+    },
   },
   plugins: [],
 }
